+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [1.3.0] 2019-04-23
+## [1.3.0] 2019-05-03
 ### Bug fixing
 - Renamed `src/layouts/Dashboard/Dashboard.jsx` to `src/layouts/Admin.jsx`
 - Renamed `src/views/Dashboard/Dashboard.jsx` to `src/views/Dashboard.jsx`